backend/requirements.txt
Flask
Flask-CORS
moviepy
openai
mysql-connector-python
connectdb
azure-cosmos
pyjwt
dotenv
google-auth-oauthlib
google-auth